+++ src/toisto/command/show_progress.py
@@ -1,5 +1,7 @@
 """Command to show progress information."""
 
+from datetime import datetime
+
 from rich.table import Table
 
 from toisto.metadata import Language, SUPPORTED_LANGUAGES
@@ -16,16 +18,19 @@ def show_progress(language: Language, quizzes: Quizzes, progress: Progress) -> N
     table.add_column("To")
     table.add_column("Answer(s)")
     table.add_column("Streak", justify="right")
+    table.add_column("Silenced until")
     sorted_quizzes = sorted(quizzes, key=lambda quiz: progress.get_progress(quiz).count, reverse=True)
     for quiz in sorted_quizzes:
         quiz_progress = progress.get_progress(quiz)
+        silence = quiz_progress.silence_until
         table.add_row(
             quiz.quiz_type.capitalize(),
             quiz.question,
             quiz.question_language,
             quiz.answer_language,
             "\n".join(quiz.answers),
-            str(quiz_progress.count) if quiz_progress.count > 0 else ""
+            str(quiz_progress.count) if quiz_progress.count > 0 else "",
+            silence.isoformat(sep=" ", timespec="minutes") if silence and silence > datetime.now() else ""
         )
     with console.pager():
         console.print(table)